Move unit test under right file
