+++ src/test/js/media_manager.test.js
@@ -3,14 +3,14 @@ import { WebRTCAdaptor } from '../../main/js/webrtc_adaptor.js';
 import { MediaManager } from '../../main/js/media_manager.js';
 
 describe("MediaManager", function() {
-	
-	
+
+
 	it("initLocalStreamWithAudio", async function() {
 
 		var adaptor = new WebRTCAdaptor({
 			 websocketURL: "ws://example.com",
 		});
-		
+
 		var mediaManager = new MediaManager({
             userParameters: {
 				mediaConstraints: {
@@ -30,22 +30,22 @@ describe("MediaManager", function() {
                 return adaptor.getSender(streamId, type)
             },
         });
-        
+
         await mediaManager.initLocalStream();
-        
+
         expect(mediaManager.localStream.getAudioTracks().length).to.be.equal(1);
         expect(mediaManager.localStream.getVideoTracks().length).to.be.equal(0);
 
-		
-		
+
+
 	});
-	
-	
+
+
 	it("getBlackVideoTrack", async function(){
 		var adaptor = new WebRTCAdaptor({
 			 websocketURL: "ws://example.com",
 		});
-		
+
 		var mediaManager = new MediaManager({
             userParameters: {
 				mediaConstraints: {
@@ -65,30 +65,30 @@ describe("MediaManager", function() {
                 return adaptor.getSender(streamId, type)
             },
         });
-       
-        
+
+
         expect(mediaManager.replacementStream).to.be.null;
         expect(mediaManager.blackVideoTrack).to.be.null;
         expect(mediaManager.blackFrameTimer).to.be.null;
-        
+
         mediaManager.getBlackVideoTrack();
-        
-        
+
+
         expect(mediaManager.replacementStream).to.not.be.null;
         expect(mediaManager.blackVideoTrack).to.not.be.null;
         expect(mediaManager.blackFrameTimer).to.not.be.null;
-        
-        
+
+
         mediaManager.stopBlackVideoTrack();
         mediaManager.clearBlackVideoTrackTimer();
         expect(mediaManager.blackVideoTrack).to.be.null;
         expect(mediaManager.blackFrameTimer).to.be.null;
-        	
+
 	});
-	
-	
+
+
 	it("turnOnOffLocalCamera", async function(){
-		
+
 		var adaptor = new WebRTCAdaptor({
 			 websocketURL: "ws://example.com",
 			 initializeComponents:false,
@@ -97,24 +97,24 @@ describe("MediaManager", function() {
 					audio:true
 			}
 		});
-		
-              
+
+
         await adaptor.mediaManager.initLocalStream();
-                
+
         expect(adaptor.mediaManager.localStream.getAudioTracks().length).to.be.equal(1);
         expect(adaptor.mediaManager.localStream.getVideoTracks().length).to.be.equal(1);
-        
+
         expect(adaptor.mediaManager.blackVideoTrack).to.be.null;
         await adaptor.mediaManager.turnOffLocalCamera();
-        
+
         expect(adaptor.mediaManager.blackVideoTrack).to.not.be.null;
-        
+
         await adaptor.turnOnLocalCamera();
         expect(adaptor.mediaManager.blackVideoTrack).to.be.null;
         expect(adaptor.mediaManager.blackFrameTimer).to.be.null;
-      	
+
 	});
-	
+
     it("testOnEndedCallback", function(done){
         this.timeout(5000);
 
@@ -139,12 +139,12 @@ describe("MediaManager", function() {
         });
 	});
     it("changeLocalVideo", async function(){
-		
+
         var videoElement = document.createElement("video");
         videoElement.id="oldElement";
         var newVideoElement = document.createElement("video");
         newVideoElement.id="newElement"
-        
+
 		var adaptor = new WebRTCAdaptor({
 			 websocketURL: "ws://example.com",
              localVideoElement : videoElement,
@@ -154,7 +154,7 @@ describe("MediaManager", function() {
 					audio:true
 			}
 		});
-		  
+
         await adaptor.mediaManager.initLocalStream();
         adaptor.mediaManager.changeLocalVideo(newVideoElement);
         expect(adaptor.mediaManager.localVideo.id).to.be.equal(newVideoElement.id);
@@ -174,7 +174,7 @@ describe("MediaManager", function() {
 					audio:true
 			}
 		});
-		  
+
         await adaptor.mediaManager.initLocalStream();
         adaptor.muteLocalMic();
         expect(adaptor.mediaManager.isMuted).to.be.equal(true);
@@ -197,7 +197,7 @@ describe("MediaManager", function() {
 					audio:true
 			}
 		});
-		  
+
         await adaptor.mediaManager.initLocalStream();
         adaptor.mediaManager.secondaryAudioTrackGainNode = adaptor.mediaManager.audioContext.createGain();
 
@@ -208,5 +208,45 @@ describe("MediaManager", function() {
 
 
 	});
-    
+
+
+  describe("checkAndStopLocalVideoTrackOnAndroid", function() {
+
+    let mediaManager;
+    let mockLocalStream;
+
+    beforeEach(function () {
+      window.isAndroid = () => {};
+
+      mockLocalStream = {
+        getVideoTracks: sinon.stub()
+      };
+
+      mediaManager = new MediaManager({
+        websocketURL: "ws://example.com",
+        initializeComponents: false,
+        localStream: mockLocalStream
+      });
+    });
+
+    it("should not stop video track if local stream exists and is not Android", function() {
+      const mockVideoTrack = { stop: sinon.fake() };
+      mockLocalStream.getVideoTracks.returns([mockVideoTrack]);
+      sinon.stub(window, 'isAndroid').returns(false);
+
+      mediaManager.checkAndStopLocalVideoTrackOnAndroid();
+
+      sinon.assert.notCalled(mockVideoTrack.stop);
+    });
+
+    it("should not stop video track if local stream does not exist", function() {
+      mediaManager.localStream = null;
+
+      mediaManager.checkAndStopLocalVideoTrackOnAndroid();
+
+      sinon.assert.notCalled(mockLocalStream.getVideoTracks);
+    });
+
+  });
+
 });
